requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==1.0.0
marshmallow==3.21.3
PyMySQL==1.1.1
cryptography==42.0.8
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==7.4.3
pytest-mock==3.12.0
requests-mock==1.11.0